+++ src/physics/narrowphase.ts
@@ -46,24 +46,6 @@ needs seperate: rotation and translation non-intersection phases.
 
 export function registerNarrowPhaseSystems(em: EntityManager) {
   // TODO(@darzu):
-  return;
-
-  console.log("NARROW PHASE DBG");
-
-  em.registerOneShotSystem(null, [AssetsDef, LocalPlayerDef], (_, res) => {
-    const b1 = em.newEntity();
-    const m1 = cloneMesh(res.assets.cube.mesh);
-    em.ensureComponentOn(b1, RenderableConstructDef, m1);
-    em.ensureComponentOn(b1, ColorDef, [0.1, 0.2, 0.1]);
-    em.ensureComponentOn(b1, PositionDef, [0, 0, 0]);
-
-    const b2 = em.newEntity();
-    const m2 = cloneMesh(res.assets.cube.mesh);
-    em.ensureComponentOn(b2, RenderableConstructDef, m2);
-    em.ensureComponentOn(b2, ColorDef, [0.1, 0.1, 0.2]);
-    em.ensureComponentOn(b2, PositionDef, [0, 0, 0]);
-    em.ensureComponentOn(b2, PhysicsParentDef, res.localPlayer.playerId);
-  });
 }
 
 type ObjWith<C extends Collider> = PhysicsObject & { collider: C };
